Created structure to load builds from dedicated files
