+++ data/builds/3-sword-style-build.ts
@@ -1,0 +1,53 @@
+import { Build } from "../builds";
+
+export const threeSwordStyleBuild: Build = {
+  slug: "3-sword-style-build",
+  title: "3 Sword Style Build",
+  slogan: "",
+  description: [
+    `Three Sword Style! Become Roranora Zoro and wield 3 Melee weapons! The Amendment with its bayonet, the Sword, and the Flag! This is a melee build, and a variation of the Spartan Build using the new Flag. While I do wish the Flag did more damage to fit the role of Med / Heavy killer better, it does allow us to bring more then one Melee weapon, and here is why you would want to do that. The Amendment helps us with any flying units that are out of reach of our blades. Shoot down Shriekers or Flying Overseers with it. If any chaff get in your way while trying to shoot, you poke their face off with that bayonet. The sword is our chaff clear. Got packs of Voteless or Hunters? Cleave them down with ease with the Sword. The Flag is our Medium / Heavy killer. Since it has a stun and an easier time hitting weakpoints we use this for the bigger guys. Overseers 3 shots, Harvesters 6 (stab its feet), Chargers 5(stab its leg), Alpha Commanders 2 (stab its head), Impalers 11 (stab its face), Tanks 10 (to the engine between the treads), Hulks 6 (to the vent). To keep us safe we will be bringing the Ballistic Shield. It blocks all melee as well as all shots that aren't rockets. For our armor we will be bringing Peak Physique for that 100% melee damage. Grenades are flex, but the new Pyro Grenades provide good crowd clear, area denial, and good damage to chunky enemies. The last 2 Strategem slots are also flex and change depending on the enemy you fight. I've found 500kg and Orbital Rail Cannon good on Bugs. Both take out Titans (your only weakness) and the 500 helps close nests for you. On Illuminate I've found Gas Mines great at shutting down breaches, the last slot is completely flex for them. You can bring Eagle Strafe to shoot down the Stingrays or clear chaff, you could bring 500kg for the packs of Meatbags, you could bring Orbital Gas as more utility. Our weakness of the Illuminate front is Meatbags. Bring something for them. On the Bot front, 110's are amazing for most heavies, and I like to bring a Hoverpack so you can drop your shield when you see a Factory Strider and go fly up on its back. Ride it like a bull and stab its back 37 times to bring it down in a blaze of Flag waving glory!`,
+  ],
+  weakness: `This build handles most things on most fronts. You will struggle if your teammates don't care about you or pay attention to your position. I am frequently knocked down by Randoms AOE, or have Strategems dropped right next to me. On Illuminate your biggest problem is Meatbags, and you probably want Strafe for the Stingray. On Bugs you can't kill Bile Titans without the ORC or 500, so if there are a bunch of them it gets rough. Bots have a lot of rockets that knock you down, so you have to be strategic and careful about when and how you engage.`,
+  image: "/assets/builds/3-sword-style-build.jpg",
+  tags: [],
+  loadout: [
+    {
+      assetId: "PH-9",
+      note: "Peak Physique armor is used for the 100% Melee damage and is required for this build",
+    },
+    {
+      assetId: "R-2",
+      note: "The Amendment is used to shoot down flying units and stab anything that gets in the way while doing so",
+    },
+    {
+      assetId: "CQC-2",
+      note: "The Sword is used to cleave through chaff like Hunters or Voteless",
+    },
+    {
+      assetId: "G-142",
+      note: "The Pyro Grenades are great at chaff clear, area denial, and can put in some work on beefy units",
+    },
+    {
+      assetId: "CQC-1",
+      note: "The Flag is our Medium / Heavy killer. We use its stun and precision to lock down and kill anything bigger than the Chaff.",
+    },
+    {
+      assetId: "SH-20",
+      note: "Ballistic Shield is used to block melee and ranged attacks to allow us to get in close to chop and stab things.",
+    },
+    {
+      assetId: "E500",
+      note: "This slot is Flex. I choose 500kg for bugs, Strafe for Illuminate and 110's for Bots.",
+    },
+    {
+      assetId: "ORS",
+      note: "This slot is also Flex. I choose Orbital Rail Cannon for bugs (for BTs), Gas Mines for Illuminate (for anti drop), and Hover Pack for bots so I can ride Factory Striders like a bull.",
+    },
+  ],
+  links: {
+    reddit: "https://www.reddit.com/r/HellDads/comments/1kp6vym/3_sword_style_build/",
+    full: "https://www.youtube.com/watch?v=mLN8A0iF5SU",
+    highlights: "https://www.youtube.com/watch?v=bcICEx0Txok",
+    discord: "https://discord.com/channels/1301285072896266252/1373449307658780723",
+  },
+};
